Fix the only 2 warnings appearing on msvc builds

+++ sqmodules/path.cpp
@@ -18,7 +18,7 @@
 void dd_simplify_fname_c(char *s)
 {
   if(!s) return;
-  int i,len=strlen(s);
+  int i,len=(int)strlen(s);
 
   // check for URL format to prevent removal of ://
   if (char *semi = (char*)memchr(s, ':', len > 8 ? 8 : len))
@@ -122,7 +122,7 @@ bool dd_fname_equal(const char *fn1,const char *fn2)
 void dd_append_slash_c(char *fn)
 {
   if(!fn) return;
-  int l=strlen(fn);
+  int l=(int)strlen(fn);
   if ( l > 0 )
     if ( fn[l-1] != PATH_DELIM_BACK && fn[l-1] != PATH_DELIM ) {
       fn[l]=PATH_DELIM;
